requirements.txt
aiogram>=3.4.0
openai>=1.12.0
python-dotenv>=1.0.0
python-docx>=1.1.0
PyPDF2>=3.0.0
PyMuPDF>=1.24.0
aiofiles>=23.0.0 